\name{spatialquantile-package}
\alias{spatialquantile-package}
\alias{spatialquantile}
\docType{package}
\title{
\packageTitle{spatialquantile}
}
\description{
\packageDescription{spatialquantile}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{spatialquantile}
\packageIndices{spatialquantile}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{spatialquantile}

Maintainer: \packageMaintainer{spatialquantile}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
